refactor: rework cert package
